implementation/requirements.txt
# Customer Support Intelligence System - LLaMA Integration
# Compatible with Anaconda/Miniconda environments

# Core Data Science
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2

# Jupyter Environment  
jupyter==1.0.0
ipykernel==6.25.0
notebook>=6.4.0

# Machine Learning & AI
torch>=2.0.1
transformers>=4.33.2
accelerate>=0.21.0
bitsandbytes>=0.41.0

# LLaMA Integration
llama-cpp-python>=0.2.11
sentence-transformers==2.2.2

# Hugging Face Integration
datasets>=2.14.4
huggingface-hub>=0.16.4
tokenizers>=0.13.3

# Text Processing
nltk==3.8.1
textblob==0.17.1

# Visualization
plotly==5.16.1
wordcloud==1.9.2

# Utilities
requests>=2.31.0
python-dotenv>=1.0.0
tqdm>=4.66.1
psutil>=5.9.0

# Optional (for advanced features)
gradio>=3.40.0
streamlit>=1.25.0